+++ src/SeaIceMomentumEquations/SeaIceMomentumEquations.jl
@@ -1,5 +1,8 @@
 module SeaIceMomentumEquations
 
+# The only functions provided by the module
+export compute_momentum_tendencies!, step_momentum!
+
 using ClimaSeaIce
 
 using Oceananigans
@@ -9,8 +12,7 @@ using Oceananigans.Operators
 using Oceananigans.Grids
 using Oceananigans.Grids: architecture
 
-# Only two functions we need to extend in this module
-import ClimaSeaIce: compute_momentum_tendencies!, step_momentum!
+using ClimaSeaIce.Rheologies: ∂ⱼ_σ₁ⱼ, ∂ⱼ_σ₂ⱼ, required_auxiliary_fields
 
 ## A Framework to solve for the ice momentum equation, in the form:
 ##
