+++ src/SeaIceMomentumEquations/explicit_momentum_equations.jl
@@ -1,4 +1,5 @@
 using Oceananigans.Utils
+using ClimaSeaIce: timestepping_coefficients
 
 const ExplicitMomentumEquation = SeaIceMomentumEquation{<:ExplicitSolver}
