+++ src/SeaIceMomentumEquations/nothing_dynamics.jl
@@ -6,3 +6,4 @@ step_momentum!(model, ::Nothing, Δt, stage) = nothing
 
 # Fallback for nothing dynamics
 compute_momentum_tendencies!(model, ::Nothing) = nothing
+
