+++ src/sea_ice_ab2_time_stepping.jl
@@ -1,4 +1,4 @@
-using ClimaSeaIce.SeaIceDynamics: step_momentum!
+using ClimaSeaIce.SeaIceMomentumEquations: step_momentum!
 
 const AB2SeaIceModel = SeaIceModel{<:Any, <:Any, <:Any, <:QuasiAdamsBashforth2TimeStepper}
 
@@ -49,12 +49,4 @@ function time_step!(model::AB2SeaIceModel, Δt; euler=false, callbacks = [])
     update_state!(model)
 
     return nothing
-end
-
-function timestepping_coefficients(ts::QuasiAdamsBashforth2TimeStepper, args...) 
-    χ  = ts.χ
-    FT = eltype(χ)
-    α  = + convert(FT, 1.5) + χ
-    β  = - convert(FT, 0.5) + χ
-    return α, β
 end
