+++ src/sea_ice_rk3_time_stepping.jl
@@ -63,13 +63,3 @@ function time_step!(model::RK3SeaIceModel, Δt; callbacks = [])
 
     return nothing
 end
-
-function timestepping_coefficients(ts::RungeKutta3TimeStepper, substep) 
-    if substep == 1
-        return ts.γ¹, zero(ts.γ¹)
-    elseif substep == 2
-        return ts.γ², ts.ζ²
-    elseif substep == 3
-        return ts.γ³, ts.ζ³
-    end
-end
